backend/data_ingestion/requirements.txt
pandas==2.1.0
elasticsearch==8.10.0
sentence-transformers==2.2.2
tqdm==4.66.1
python-dotenv==1.0.0 